pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "attachments"
version = "0.6.0"
description = "A modular file processing pipeline with automatic type dispatch"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.8"
dependencies = []

[project.optional-dependencies]
full = [
    "pandas>=1.0.0",
    "pdfplumber>=0.7.0", 
    "python-pptx>=0.6.0",
    "Pillow>=8.0.0",
    "requests>=2.25.0",
    "beautifulsoup4>=4.9.0"
]
dev = [
    "pytest>=7.0.0"
]

[project.urls]
Homepage = "https://github.com/yourusername/attachments"
Repository = "https://github.com/yourusername/attachments"

[tool.setuptools]
packages = ["attachments"] 